backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
drf-spectacular==0.27.2
django-cors-headers==4.3.1
